do not use 'long' to store 4 bytes, as 64bit architectures have 64bit longs.

+++ lib/md5.c
@@ -53,7 +53,7 @@ documentation and/or software.
 #include <string.h>
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;
 
 /* MD5 context. */
 struct md5_ctx {
